Removing the legacy mailing lists (#204)
These mailing lists were primarily used for okd 3.x.  There has been no
new threads since May and only 1 on the dev list this year and maybe 3
on the user list.

I intend to send a final email to these lists once this merges and is
live on the site letting people know that stackoverflow is the
appropriate place for community 3.11 questions and to join the working
group or slack channel.